chore: relocate tests
