+++ src/test/java/com/example/Contractor/Controller/CountryControllerTest.java
@@ -1,6 +1,5 @@
-package com.example.Contractor;
+package com.example.Contractor.Controller;
 
-import com.example.Contractor.Controller.CountryController;
 import com.example.Contractor.DTO.Country;
 import com.example.Contractor.Service.CountryService;
 import org.junit.jupiter.api.Assertions;
