+++ src/test/java/com/example/Contractor/UIController/ContractorUIControllerTest.java
@@ -1,5 +1,6 @@
-package com.example.Contractor;
+package com.example.Contractor.UIController;
 
+import com.example.Contractor.DatabaseSetup;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
